refactor: split opencode parser module
